Extract base validation class working against a feature type, but not requiring a FeatureTypeInfo to be present
